Add kernel counting logic

+++ src/tools/simeng/main.cc
@@ -10,12 +10,14 @@
 #include "simeng/version.hh"
 
 /** Tick the provided core model until it halts. */
-int simulate(simeng::Core& core, simeng::MemoryInterface& dataMemory,
-             simeng::MemoryInterface& instructionMemory) {
+uint64_t simulate(simeng::Core& core, simeng::MemoryInterface& dataMemory,
+                  simeng::MemoryInterface& instructionMemory) {
   uint64_t iterations = 0;
 
   // Tick the core and memory interfaces until the program has halted
   while (!core.hasHalted() || dataMemory.hasPendingRequests()) {
+    //    std::cerr << "tick" << std::endl;
+
     // Tick the core
     core.tick();
 
@@ -91,7 +93,7 @@ int main(int argc, char** argv) {
 
   // Run simulation
   std::cout << "[SimEng] Starting...\n" << std::endl;
-  int iterations = 0;
+  uint64_t iterations = 0;
   auto startTime = std::chrono::high_resolution_clock::now();
   iterations = simulate(*core, *dataMemory, *instructionMemory);
